refactor(ClassroomMonitorModule): Extract DataExportModule
#1225
